Ui wlth blac and data  movie detail screen

+++ lib/movies/data/model/movie_detalis_model.dart
@@ -24,5 +24,5 @@ class MovieDetailsModel extends MovieDetail {
           releaseDate: json['release_date'],
           runtime: json['runtime'],
           title: json['title'],
-          voteAverage: json['vote_average']);
+          voteAverage: json['vote_average'].toDouble());
 }
